Routing Errors Fixed on All Pages
Since application.html.erb seems to be doing a good job at orchestrating
the pick-up of the .css and .js files that we need, it is unnecessary,
and results in errors, to try and include them again. This commit aims
to remedy that by removing all such attempts to do so!